added links to all menus
