+++ UnboundLib/UnboundLib.csproj
@@ -44,6 +44,7 @@
   <ItemGroup>
     <None Remove="UnboundLib.csproj.user.dist" />
     <None Remove="PostBuild.msbuild" />
+    <None Remove="unboundlinks" />
     <None Include="..\README.md" />
     <None Include="..\.gitignore" />
   </ItemGroup>
@@ -52,6 +53,7 @@
     <EmbeddedResource Include="..\AssetsEmbedded\unboundui" />
     <EmbeddedResource Include="..\AssetsEmbedded\modoptionsui" />
     <EmbeddedResource Include="..\AssetsEmbedded\toggle ui" />
+    <EmbeddedResource Include="..\AssetsEmbedded\unboundlinks" />
   </ItemGroup>
 
   <ItemGroup>
